feat(typeorm): add typed entity column protection
Add an ability to specify custom type for protected columns.

This allows more conveniently to work with data and decrypt it, only
when it's needed. Typeorm transformer just initiates a wrapper type
and nothing more.

All others operations(decryption and etc.) will be done, when someone requests
this data for the first. In subsequent times decrypted value will be retrieved
from cache.